+++ src/SqlArtisan/Internal/SqlBuilder/Delete/DeleteBuilder.cs
@@ -11,6 +11,9 @@ internal class DeleteBuilder(params SqlPart[] rootParts) :
     public SqlStatement Build(Dbms dbms) =>
         BuildCore(dbms);
 
+    public IReturningBuilder Returning(params object[] expressions) =>
+        ReturningBuilder.Create(this, expressions);
+
     public IDeleteBuilderWhere Where(SqlCondition condition)
     {
         AddPart(new WhereClause(condition));