+++ src/SqlArtisan/Internal/SqlBuilder/Delete/IDeleteBuilderDelete.cs
@@ -1,6 +1,6 @@
 ﻿namespace SqlArtisan.Internal;
 
-public interface IDeleteBuilderDelete : ISqlBuilder
+public interface IDeleteBuilderDelete : ISqlBuilder, IReturning
 {
     IDeleteBuilderWhere Where(SqlCondition condition);
 }