+++ src/SqlArtisan/Internal/SqlBuilder/SqlBuilderBase.cs
@@ -4,7 +4,23 @@ internal abstract class SqlBuilderBase(SqlPart[] rootParts)
 {
     private readonly List<SqlPart> _parts = [.. rootParts];
 
-    protected void AddPart(SqlPart part) => _parts.Add(part);
+    protected internal void AddPart(SqlPart part) => _parts.Add(part);
+
+    internal SqlStatement BuildWithPart(SqlPart extraPart, Dbms dbms)
+    {
+        _parts.Add(extraPart);
+        try
+        {
+            return BuildCore(dbms);
+        }
+        finally
+        {
+            _parts.RemoveAt(_parts.Count - 1);
+        }
+    }
+
+    internal SqlStatement BuildWithPart(SqlPart extraPart) =>
+        BuildWithPart(extraPart, SqlArtisanConfig.DefaultDbms);
 
     protected SqlStatement BuildCore(Dbms dbms)
     {