+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - Added overloads to the `Decode` method to remove the array requirement. (#39)
+- Added support for RETURNING clause. (#41)
 
 ## [0.2.0-beta.2] - 2025-07-19
 ### Added